+++ app/javascript/lca/components/characterEditor/editors/SiderealExaltEditor.tsx
@@ -1,4 +1,3 @@
-// @flow
 import React from 'react'
 
 import Typography from '@material-ui/core/Typography'
@@ -10,8 +9,8 @@ import BlockPaper from 'components/generic/blockPaper.jsx'
 import { nonCasteAbilities } from 'utils/calculated'
 import type { Character } from 'utils/flow-types'
 
-type Props = { character: Character, onChange: Function }
-function DragonbloodExaltEditor({ character, onChange }: Props) {
+type Props = { character: Character; onChange: React.ChangeEventHandler }
+function SiderealExaltEditor({ character, onChange }: Props) {
   let caste_abilities = character.caste_abilities || []
   if (character.caste === 'water') {
     caste_abilities = [...caste_abilities, 'martial arts'].sort()
@@ -19,10 +18,8 @@ function DragonbloodExaltEditor({ character, onChange }: Props) {
 
   return (
     <BlockPaper>
-      <DbAspectSelect value={character.caste} onChange={onChange} />
-
       <Typography style={{ marginTop: '0.5em', textTransform: 'capitalize' }}>
-        Aspect Abilities: {caste_abilities.join(', ')}
+        Caste Abilities: {caste_abilities.join(', ').replace('_', ' ')}
       </Typography>
 
       <AbilitySelect
@@ -38,4 +35,4 @@ function DragonbloodExaltEditor({ character, onChange }: Props) {
     </BlockPaper>
   )
 }
-export default DragonbloodExaltEditor
+export default SiderealExaltEditor