+++ app/javascript/lca/components/characterEditor/exaltTraits/SolarCasteSelect.tsx
@@ -4,21 +4,22 @@ import ListSubheader from '@material-ui/core/ListSubheader'
 import MenuItem from '@material-ui/core/MenuItem'
 import TextField, { TextFieldProps } from '@material-ui/core/TextField'
 
-const LunarCasteSelect = (props: TextFieldProps) => (
+const SolarCasteSelect = (props: TextFieldProps) => (
   <TextField
     select
     {...props}
-    data-cy="select-db-aspect"
+    name="caste"
     label="Caste"
     margin="dense"
-    name="caste"
+    data-cy="select-solar-caste"
   >
     <ListSubheader key="none">Select a Caste</ListSubheader>
-    <MenuItem value="full moon">Full Moon</MenuItem>
-    <MenuItem value="changing moon">Changing Moon</MenuItem>
-    <MenuItem value="no moon">No Moon</MenuItem>
-    <MenuItem value="casteless">Casteless</MenuItem>
+    <MenuItem value="dawn">Dawn</MenuItem>
+    <MenuItem value="zenith">Zenith</MenuItem>
+    <MenuItem value="twilight">Twilight</MenuItem>
+    <MenuItem value="night">Night</MenuItem>
+    <MenuItem value="eclipse">Eclipse</MenuItem>
   </TextField>
 )
 
-export default LunarCasteSelect
+export default SolarCasteSelect