+++ app/javascript/lca/components/characterEditor/index.jsx
@@ -17,6 +17,7 @@ import CustomAttributeExaltEditor from './editors/customAttributeExaltEditor.jsx
 import CustomEssenceExaltEditor from './editors/customEssenceExaltEditor.jsx'
 import DragonbloodExaltEditor from './editors/DragonbloodExaltEditor.jsx'
 import LunarExaltEditor from './exaltTraits/LunarExaltEditor'
+import SiderealExaltEditor from './editors/SiderealExaltEditor'
 import HealthLevelEditor from './editors/healthLevelEditor.jsx'
 import IntimacyEditor from './editors/intimacyEditor.jsx'
 import LimitEditor from './editors/limitEditor.jsx'
@@ -111,6 +112,11 @@ class CharacterEditor extends Component<Props> {
           {character.type === 'LunarCharacter' && (
             <Grid item xs={12} md={6}>
               <LunarExaltEditor character={character} onChange={handleChange} />
+            </Grid>
+          )}
+          {character.type === 'SiderealCharacter' && (
+            <Grid item xs={12} md={6}>
+              <SiderealExaltEditor character={character} onChange={handleChange} />
             </Grid>
           )}
           {character.type === 'CustomAbilityCharacter' && (