+++ spec/models/custom_ability_character_spec.rb
@@ -5,7 +5,7 @@ require 'rails_helper'
 RSpec.describe CustomAbilityCharacter, type: :model do
   describe 'converting types' do
     %i[
-      character solar_character dragonblood_character lunar_character
+      character solar_character dragonblood_character lunar_character sidereal_character
       custom_attribute_character custom_essence_character
     ].each do |char|
       it "works for #{char}" do